+++ utils/number_test.go
@@ -3,12 +3,12 @@ package utils
 import "testing"
 
 func TestParseInteger(t *testing.T) {
-	integer, err := ParseInteger("123")
+	integer, err := ParseInt64("123")
 	if err != nil {
 		t.Fatal(err)
 	}
 	t.Log(integer)
-	integer, err = ParseInteger("")
+	integer, err = ParseInt64("")
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -16,12 +16,12 @@ func TestParseInteger(t *testing.T) {
 }
 
 func TestParseDecimal(t *testing.T) {
-	decimal, err := ParseDecimal("123.1")
+	decimal, err := ParseFloat64("123.1")
 	if err != nil {
 		t.Fatal(err)
 	}
 	t.Log(decimal)
-	decimal, err = ParseDecimal("")
+	decimal, err = ParseFloat64("")
 	if err != nil {
 		t.Fatal(err)
 	}
